Remove unused references to JsonProvider.
